+++ CHANGELOG.md
@@ -1,13 +1,27 @@
 ### Changelog
 
-## [0.1.1] - 11/24/2018
+# 0.1.2 - 11/24/2018
+* Added
+    ### Properties:
+    | Property   | Default |
+    |------------|:-------:|
+    | borderRadius  | 5.0 |
+    | keybaordType  | number |
+    | keyboardAction  | next |
+
+# 0.1.1 - 11/24/2018
 
 * Added
-    Example
-*   Properties 
-        * fontSize
-        * isTextObscure
+    ### Properties:
+    | Property   | Default |
+    |------------|:-------:|
+    | onSubmit  | Function |
+    | fieldsCount  | 4 |
+    | isTextObscure  | false |
+    | fontSize  | 40.0 |
 
-## [0.0.1] - 10/24/2018
+
+
+# 0.0.1 - 10/24/2018
 
 * Initial release, working base functionality